Upload

+++ 2021.02 - [ Entity Framework Core ]/06 - [ EF Introduction - Exercise ]/01 - [ Solutions ]/14 - [ Delete Project by Id ]/StartUp.cs
@@ -1,13 +1,15 @@
 ﻿namespace SoftUni
 {
-    using Microsoft.EntityFrameworkCore;
     using SoftUni.Data;
+
     using System;
     using System.Linq;
     using System.Text;
 
     public class StartUp
     {
+        private const int ProjectId = 2;
+
         static void Main(string[] args)
         {
             using var context = new SoftUniContext();
@@ -19,18 +21,20 @@
         {
             var project2 = context
                 .Projects
-                .FirstOrDefault(x => x.ProjectId == 2);
+                .FirstOrDefault(x => x.ProjectId == ProjectId);
 
-            foreach (var entity in context.EmployeesProjects.Where(x => x.ProjectId == 2))
+            foreach (var entity in context.EmployeesProjects.Where(x => x.ProjectId == ProjectId))
             {
                 context.EmployeesProjects.Remove(entity);
             }
 
             context.SaveChanges();
 
-            context.Projects.Remove(project2);
-
-            context.SaveChanges();
+            if (project2 != null)
+            {
+                context.Projects.Remove(project2);
+                context.SaveChanges();
+            }
 
             StringBuilder sb = new StringBuilder();
