+++ 2021.02 - [ Entity Framework Core ]/06 - [ EF Introduction - Exercise ]/01 - [ Solutions ]/15 - [ Remove Town ]/StartUp.cs
@@ -1,13 +1,16 @@
 ﻿namespace SoftUni
 {
-    using Microsoft.EntityFrameworkCore;
     using SoftUni.Data;
+
+    using Microsoft.EntityFrameworkCore;
     using System;
     using System.Linq;
     using System.Text;
 
     public class StartUp
     {
+        private const string TownName = "Seattle";
+
         static void Main(string[] args)
         {
             using var context = new SoftUniContext();
@@ -19,7 +22,7 @@
         {
             var town = context
                 .Towns
-                .FirstOrDefault(x => x.Name == "Seattle");
+                .FirstOrDefault(x => x.Name == TownName);
 
             var addresses = context
                 .Addresses
@@ -27,7 +30,6 @@
                 .Where(x => x.Town == town)
                 .ToList();
 
-            var counter = 0;
             foreach (var entity in addresses)
             {
                 foreach (var employee in entity.Employees.Where(x => x.AddressId == entity.AddressId))
@@ -36,15 +38,17 @@
                 }
 
                 context.Addresses.Remove(entity);
-                counter++;
             }
 
-            context.SaveChanges();
-            context.Towns.Remove(town);
-            context.SaveChanges();
+            if (town != null)
+            {
+                context.SaveChanges();
+                context.Towns.Remove(town);
+                context.SaveChanges();
+            }
 
             StringBuilder sb = new StringBuilder();
-            sb.AppendLine($"{counter} addresses in Seattle were deleted");
+            sb.AppendLine($"{addresses.Count} addresses in Seattle were deleted");
             return sb.ToString().TrimEnd();
         }
     }
